requirements.txt
robotframework==3.2.2
selenium==3.141.0
python-dateutil==2.8.1
pytz==2019.3
regex==2020.2.20
setuptools==18.2
robotframework-seleniumlibrary==4.4.0
robotframework-databaselibrary==1.2.4
cx_oracle==8.0.0
robotframework-jsonlibrary==0.3.1
jsonpath==0.82
robotframework-requests==0.9.1
selenium==3.141.0
requests==2.27.1

 


